+++ src/ist/meic/bomberman/engine/Robot.java
@@ -5,7 +5,7 @@ import android.content.Context;
 
 public class Robot extends Entity {
 
-	public Robot(Context c, double dyingPoints, int x, int y) {
+	public Robot(Context c, int dyingPoints, int x, int y) {
 		super(c, R.drawable.robot, dyingPoints, x, y);
 	}
